Update package.json scripts
