+++ apps/client/src/app.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get, HttpCode } from '@nestjs/common';
-import { Public } from 'apps/hello-gsm-back/src/auth/decorators/public.decorator';
+import { Public } from 'apps/client/src/auth/decorators/public.decorator';
 
 @Controller('')
 export class AppController {
