+++ apps/client/src/application/dto/application.detail.qualification.dto.ts
@@ -1,5 +1,5 @@
 import { IsEnum, IsString } from 'class-validator';
-import { EducationStatus } from 'apps/hello-gsm-back/src/types';
+import { EducationStatus } from 'apps/client/src/types';
 import { ApplicationDetailSuperDto } from './application.detail.super';
 
 export class ApplicationDetailQualificationDto extends ApplicationDetailSuperDto {
