+++ apps/client/src/application/dto/seconds.submission.dto.ts
