+++ apps/client/src/auth/auth.controller.ts
@@ -17,8 +17,8 @@ import {
   ApiTags,
 } from '@nestjs/swagger';
 import { Request, Response } from 'express';
-import { ENV } from 'apps/hello-gsm-back/src/lib/env';
-import { AtUser } from 'apps/hello-gsm-back/src/types';
+import { ENV } from 'apps/client/src/lib/env';
+import { AtUser } from 'apps/client/src/types';
 import { AuthService } from './auth.service';
 import { Public } from './decorators/public.decorator';
 import { User } from './decorators/user.decorator';
@@ -28,7 +28,7 @@ import {
   accessToken,
   refreshToken,
   registerToken,
-} from 'apps/hello-gsm-back/src/utils/token.name';
+} from 'apps/client/src/utils/token.name';
 
 @ApiTags('Auth')
 @Controller('auth')
