+++ apps/client/src/auth/dto/register.dto.ts
