+++ apps/client/src/auth/strategies/kakao.strategy.ts
@@ -2,7 +2,7 @@ import { PassportStrategy } from '@nestjs/passport';
 import { Strategy, Profile } from 'passport-kakao';
 import { Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import { ENV } from 'apps/hello-gsm-back/src/lib/env';
+import { ENV } from 'apps/client/src/lib/env';
 
 @Injectable()
 export class KakaoStrategy extends PassportStrategy(Strategy, 'kakao') {
