+++ apps/client/src/auth/types/kakao.user.type.ts
