+++ apps/client/src/lib/logger.middleware.ts
