+++ apps/client/src/types/EducationStatus.ts
