+++ apps/client/src/types/Major.ts
