+++ apps/client/test/jest-e2e.json
