+++ nest-cli.json
@@ -1,21 +1,20 @@
 {
   "$schema": "https://json.schemastore.org/nest-cli",
   "collection": "@nestjs/schematics",
-  "sourceRoot": "apps/hello-gsm-back/src",
+  "sourceRoot": "apps/client/src",
   "monorepo": true,
-  "root": "apps/hello-gsm-back",
+  "root": "apps/client",
   "compilerOptions": {
-    "webpack": true,
-    "tsConfigPath": "apps/hello-gsm-back/tsconfig.app.json"
+    "tsConfigPath": "apps/client/tsconfig.app.json"
   },
   "projects": {
-    "hello-gsm-back": {
+    "client": {
       "type": "application",
-      "root": "apps/hello-gsm-back",
+      "root": "apps/client",
       "entryFile": "main",
-      "sourceRoot": "apps/hello-gsm-back/src",
+      "sourceRoot": "apps/client/src",
       "compilerOptions": {
-        "tsConfigPath": "apps/hello-gsm-back/tsconfig.app.json"
+        "tsConfigPath": "apps/client/tsconfig.app.json"
       }
     },
     "admin": {
