+++ package.json
@@ -7,12 +7,17 @@
   "license": "UNLICENSED",
   "scripts": {
     "prebuild": "rimraf dist",
-    "build": "nest build",
+    "build": "nest build client",
+    "build:a": "nest build admin",
     "format": "prettier --write \"apps/**/*.ts\" \"libs/**/*.ts\"",
-    "start": "nest start",
-    "start:dev": "nest start --watch",
-    "start:debug": "nest start --debug --watch",
-    "start:prod": "node dist/main",
+    "start": "nest start client",
+    "start:dev": "nest start client --watch",
+    "start:debug": "nest start client --debug --watch",
+    "start:prod": "node dist/apps/client/main",
+    "start:a": "nest start admin",
+    "start:a:dev": "nest start admin --watch",
+    "start:a:debug": "nest start admin --debug --watch",
+    "start:a:prod": "node dist/apps/admin/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "test": "jest",
     "test:watch": "jest --watch",
